refactor Media/MediaList/MediaListPlayer
Prepare for new Media* class that follow libvlc API.

Rename Media to MediaWrapper since it's a wrapper of the future Media class,
and also a wrapper of MediaDatabase elements.

Move these classes from libvlc to vlc-android since these classes depends too
much of the vlc application. Indeed there are mainly used by MediaDatabase that
belongs to application and not the lib.