build: fix ODR violation
This commit fixes the regression introduced by
5a06f68fc4787628e960e6afc00588f3950e774e that switched to the c++ static
runtime and caused exceptions to be uncaught.

As https://developer.android.com/ndk/guides/cpp-support.html#static_runtimes
explains it, one process should *NEVER* link with more than one shared
libraries that link with the c++ static library.

To fix this issue: libmla and libvlc are now static and linked together as
libvlcjni.so with the c++ static library.