libvlc: fix smb2/dsm conflicts
The smb2 module is back to higher priority again. A lot of bugs has been fixed
since the first release, notably the signing security mode. This fixes
impossible login to smb2 servers, when dsm requested first a login and smb2 was
never tried.

There is also a new patch that prevent to request 2 dialogs from both smb2 and
dsm from the same input. If smb2 login failed, it is useless to request
credentials via the dsm plugin.

PS: if the module smb2 failed because it tried to connect to a smb1 server, it
will fail without requesting any credentials, therefore the dsm module will be
able to request them.