libvlc: MediaPlayer: post a surface update after vout events
At least with LibVLC 4, the video surface stays black until the
orientation of the phone is changed, calling this event again after the
creation of the VLC vout display.

It might come from a wrong ordering and reporting of parent layout
events while the inner views are not reading for display, so post an
event at the end of the current mainloop re-updating the surfaces to be
sure that the video surface is visible in LibVLC client applications.

This a hack, but it is what the VLC for Android application is currently
doing too.

Refs #1433