buildsystem: libvlc: move lua linkage hack
In libVLC 3.x, the lua contrib didn't expose a pkg-config file so it
wasn't picked by the pkg-config probing in the compile-libvlc.sh script.

The plan is to expose it on libVLC 4.x, so to keep compatibility with
both version only add the lua link library if the pkg-config file
doesn't exist.